[ios] Support selecting moving user dot
Selecting the user location annotation now works while the annotation is moving. As it moves, the annotation remains selected and the callout view tracks the annotation. There is a tricky special case, which is that the callout must point to the annotation view’s implicit frame but must quickly rendezvous with the explicit frame.